refactor(test): reorganize test configuration
Move and improve test configuration for better maintainability.

- Move TestConfig to dedicated testutils package
- Update test dependencies to use new location
- Add additional test configuration properties
