+++ src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -8,5 +8,25 @@
     "name": "app.base-url",
     "type": "java.lang.String",
     "description": "Base URL of the application"
+  },
+  {
+    "name": "test.transaction.timeout",
+    "type": "java.lang.String",
+    "description": "A description for 'test.transaction.timeout'"
+  },
+  {
+    "name": "test.concurrent.max-threads",
+    "type": "java.lang.String",
+    "description": "A description for 'test.concurrent.max-threads'"
+  },
+  {
+    "name": "test.retry.max-attempts",
+    "type": "java.lang.String",
+    "description": "A description for 'test.retry.max-attempts'"
+  },
+  {
+    "name": "test.retry.delay",
+    "type": "java.lang.String",
+    "description": "A description for 'test.retry.delay'"
   }
 ]}
